Move tests out of testresources
This aligns us with other testing-cabal projects.
