+++ tests/test_resource_graph.py
@@ -25,7 +25,7 @@ from testresources import _resource_graph
 
 
 def test_suite():
-    from testresources.tests import TestUtil
+    from . import TestUtil
 
     loader = TestUtil.TestLoader()
     result = loader.loadTestsFromName(__name__)
