+++ tests/test_test_resource.py
@@ -26,7 +26,9 @@ from testresources.tests import (
 
 
 def test_suite():
-    loader = testresources.tests.TestUtil.TestLoader()
+    from . import TestUtil
+
+    loader = TestUtil.TestLoader()
     result = loader.loadTestsFromName(__name__)
     return result
