+++ tox.ini
@@ -6,7 +6,8 @@ minversion = 4.0
 usedevelop = true
 extras =
   test
-commands = python -m testtools.run testresources.tests.test_suite
+commands =
+  python -W once -m testtools.run tests.test_suite {posargs}
 
 [testenv:ruff]
 description =
@@ -14,5 +15,5 @@ description =
 deps =
   ruff
 commands =
-  ruff format .
   ruff check --fix --unsafe-fixes --exit-non-zero-on-fix .
+  ruff format .
